requirements.txt
# King's Choice Management App Dependencies

# Core Flask Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1

# Database
SQLAlchemy==2.0.35

# Development dependencies (optional for production)
python-dotenv==1.0.0

# Production server (for deployment)
gunicorn==21.2.0
waitress==2.1.2

# Security and utilities
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7

# For better JSON handling
flask-cors==4.0.0

# Telegram Bot
python-telegram-bot==20.7
requests==2.31.0

# Translation
googletrans==4.0.0